better ui + icons + bonus structures with active bonuses + refractory code

+++ Assets/Scripts/Managers/PlayerMoneyManager.cs
@@ -7,7 +7,7 @@ using TMPro;
 public class PlayerMoneyManager : MonoBehaviour
 {
     public static PlayerMoneyManager Instance { get; private set; }
-    private int playerMoneyAmount { get; set; }
+    private float playerMoneyAmount { get; set; }
 
     PlayerUIManager playerUIManager;
 
@@ -24,12 +24,12 @@ public class PlayerMoneyManager : MonoBehaviour
         }
         playerUIManager = FindObjectOfType<PlayerUIManager>();
     }
-    public void SetAmount(int amount)
+    public void SetAmount(float amount)
     {
         playerMoneyAmount += amount;
         playerUIManager.UpdateAmountVisual(playerMoneyAmount);
     }
-    public int GetAmount()
+    public float GetAmount()
     {
         return playerMoneyAmount;
     }
